[Bugfix] Re-land MiniMax M3 default video processor
